feat: extra language inputs for codegen
When using language plugins like Cap'n'proto for Java from Bazel,
there needs to be a way to add language extras as inputs to the
codegen phase.

This is only needed when generating capnp code in multiple
languages. Even if the generator is only producing a C++ in a
given invocation, the imports for other languages will fail to be
found due to Bazel's strict sandboxing. This new attribute fixes
that condition.

Fixes and closes capnproto/capnproto#1903.

Relates-To: elide-dev/bigtent#6
Relates-To: capnproto/capnproto#1903
Signed-off-by: Sam Gammon <sam@elide.ventures>